+++ client/src/screens/Home/Home.jsx
@@ -1,0 +1,5 @@
+export default function Home() {
+  return(
+    <h1>Home</h1>
+  )
+}
