+++ client/src/screens/PostCreate/PostCreate.jsx
@@ -1,0 +1,5 @@
+export default function PostCreate() {
+  return(
+    <h1>PostCreate</h1>
+  )
+}
