+++ client/src/screens/PostDetail/PostDetail.jsx
@@ -1,0 +1,5 @@
+export default function PostDetail() {
+  return(
+    <h1>PostDetail</h1>
+  )
+}
