+++ client/src/screens/PostEdit/PostEdit.jsx
@@ -1,0 +1,5 @@
+export default function PostEdit() {
+  return(
+    <h1>PostEdit</h1>
+  )
+}
